Moved a number of openwam classes into the OWD namespace.
Changed the TrajControl class over to JointPos types instead of vanilla
  vector<double> types, since they have built-in vector math functions.

Moved link7 origin forward to a point in front of the palm of the hand, so
  that  the link definitions could be used for more reasonable workspace
  control of the end effector position.

Added a couple of helpful operators to openmath classes.

Fixed the output operator for the SO3 class to use the supplied stream.

Added Plugin functions for working with the Jacobian Pseudo-Inverse and
  also doing nullspace projections of joint motions.

Added code to allow a running 280 hand to be reset, which is not working yet.

Added a ReloadPlugins service call (not always working yet).